+++ GameEngineVS/GameEngine/GameEngine/Vector3D.cpp
@@ -11,7 +11,7 @@ Vector3D::Vector3D(float x, float y, float z, int w) : x(x), y(y), z(z), w(w)
 
 float Vector3D::Magnitude() const
 {
-	return sqrt(pow(x,2) + pow(y,2) + pow(z,2));
+	return (float) sqrt(pow(x,2) + pow(y,2) + pow(z,2));
 }
 
 void Vector3D::Normalize()
